requirements.txt
Flask
Flask-Cors
mongoengine
pymongo
bcrypt
PyJWT
qrcode
python-dotenv
